Use monotonic clock for time measurements
Currently, watchdog daemon doesn't play nice around abrupt time steps.
This might be a problem on devices without RTC,
which only obtain correct time after the startup.
By this time, watchdog daemon and the processes it monitors
may be running for some time.
However, watchdog daemon may be fooled by a time jump,
and decide to reboot the system.

So, use monotonic time for all time, duration, and timeout
measurements and calculations.

Introduce a simple wrapper around clock_gettime() to use as a drop-in
replacement for time(NULL).

As a replacement for timeradd() and timersub(),
add helper functions (inspired by BSD's macros of the same name)
for arithmetics on timespec.

Signed-off-by: Andrey Mazo <amazo@checkvideo.com>
Reviewed-by: Arun Whitchurch <awhitchurch@checkvideo.com>