Use monotonic clock for file mtime monitoring
Comparing file's mtime against current time
gives surprising results around time jumps.
So, instead of doing so,
keep track of when a file's mtime was changed
according to our monotonic clock.

Don't care about how file's mtime changes
as long as it changes.

This not only works properly around time steps,
but also allows file's mtime to go backwards if needed.

The internally recorded file's mtime according to the monotonic clock
is only as accurate as "interval" though.
E.g., if a file was changed at time=1, but interval=30,
then we'll perceive the file as if it changed at time=30.
So "change" config option may need to be larger to account for this.

Signed-off-by: Andrey Mazo <amazo@checkvideo.com>
Reviewed-by: Arun Whitchurch <awhitchurch@checkvideo.com>