Fix stale PID for watchdog service
- Patch by Arijit Banerjee for Debian bug #892715 / Launchpad #1750942

- Original explanation:

On Ubuntu 14.04, it was noticed that "sudo service watchdog start"
command was sometimes hanging indefinitely. Investigation revealed
that the watchdog init script tries sending a SIGTERM to the process
present in the pidfile for the wd_keepalive daemon. However, the
init script does not seem to clear pidfiles while stopping services,
so the pidfile is often stale. This resulted in SIGTERM being sent
not to wd_keepalive, but to a service which was not responding to
SIGTERM, causing the command mentioned to hang.

This issue seems to be present even on the latest version of the
watchdog package.

The issue was worked around by explicitly specifying the name
of the executable to stop in "start-stop-daemon --stop". The man
page for start-stop-daemon explicitly mentions that not doing
so while supplying a pidfile can cause unintended processes to
be acted upon.